requirements.txt
pydantic==2.6.0
psycopg2-binary==2.9.9
pymongo==4.6.1